+++ tango/io/vfs/FileFolder.d
@@ -362,7 +362,7 @@ class FileGroup : VfsFiles
 
         ***********************************************************************/
 
-        final uint files ()
+        final size_t files ()
         {
                 return group.length;
         }
@@ -438,9 +438,9 @@ private class FolderGroup : VfsFolders
 
         ***********************************************************************/
 
-        final uint files ()
+        final size_t files ()
         {
-                uint files;
+                size_t files;
                 foreach (folder; members)
                          files += folder.stats.files;
                 return files;
@@ -467,7 +467,7 @@ private class FolderGroup : VfsFolders
 
         ***********************************************************************/
 
-        final uint folders ()
+        final size_t folders ()
         {
                 if (members.length is 1)
                     return members[0].stats.folders;
@@ -480,7 +480,7 @@ private class FolderGroup : VfsFolders
 
         ***********************************************************************/
 
-        final uint entries ()
+        final size_t entries ()
         {
                 return files + folders;
         }
